MOBILE-108: Reorganize Example app from type-based to feature-based grouping
