--- Example/Example/Models/Payload.swift
@@ -1,14 +1,0 @@
-//
-//  Payload.swift
-//  Example
-//
-//  Created by Sergei Semko on 6/10/24.
-//  Copyright © 2024 Mindbox. All rights reserved.
-//
-
-import Foundation
-
-public struct Payload: Codable {
-    var pushName: String
-    var pushDate: String
-}
