+++ Example/Example/NotificationCenter/SwiftDataManager.swift
